pyproject.toml
[project]
name = "terrarium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.13.3",
    "dotenv>=0.9.9",
    "influxdb-client>=1.49.0",
    "pydantic>=2.12.5",
    "pyserial>=3.5",
    "requests>=2.32.5",
]
